# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "AniDB API Documentation",
        "version": "1.0"
    },
    "paths": {
...
        "/api/groups": {
            "get": {
                "tags": [
                    "Group"
                ],
                "summary": "Get All Groups",
                "operationId": "74f97965cba29ff042644d03ab580adc",
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Group"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "$ref": "#/components/responses/Unauthorized"
                    }
                },
                "security": [
                    {
                        "token": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Group"
                ],
                "summary": "Add a Group",
                "operationId": "d5e58c0df5e02a3de23c70700551aa64",
                "parameters": [
                    {
                        "name": "name",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "Sample Group Name"
                    }
                ],
                "responses": {
                    "200": {
                        "$ref": "#/components/responses/Success"
                    },
                    "401": {
                        "$ref": "#/components/responses/Unauthorized"
                    }
                },
                "security": [
                    {
                        "token": []
                    }
                ]
            }
        },
        "/api/groups/names": {
            "get": {
                "tags": [
                    "Group"
                ],
                "summary": "Get All Group Names",
                "operationId": "53fbdf770508b509a82ce8241e563ba9",
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "type": "string"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "$ref": "#/components/responses/Unauthorized"
                    }
                },
                "security": [
                    {
                        "token": []
                    }
                ]
            }
        },
        "/api/groups/{group_id}": {
            "put": {
                "tags": [
                    "Group"
                ],
                "summary": "Edit a Group",
                "operationId": "c6d4c44e7f3880c4e941de9d7f9ec40c",
                "parameters": [
                    {
                        "name": "group_id",
                        "in": "path",
                        "description": "Group ID",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        },
                        "example": "e9597119-8452-4f2b-96d8-f2b1b1d2f158"
                    },
                    {
                        "name": "name",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "Sample Group Name"
                    }
                ],
                "responses": {
                    "200": {
                        "$ref": "#/components/responses/Success"
                    },
                    "401": {
                        "$ref": "#/components/responses/Unauthorized"
                    }
                },
                "security": [
                    {
                        "token": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Group"
                ],
                "summary": "Delete a Group",
                "operationId": "89ce3046c99740e998e236b35854b8da",
                "parameters": [
                    {
                        "name": "group_id",
                        "in": "path",
                        "description": "Group ID",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        },
                        "example": "e9597119-8452-4f2b-96d8-f2b1b1d2f158"
                    }
                ],
                "responses": {
                    "200": {
                        "$ref": "#/components/responses/Success"
                    },
                    "401": {
                        "$ref": "#/components/responses/Unauthorized"
                    }
                },
                "security": [
                    {
                        "token": []
                    }
                ]
            }
        },
...
        "/api/mal/{title_id}": {
            "get": {
                "tags": [
                    "MAL"
                ],
                "summary": "Retrieve Title Information",
                "operationId": "8288a9cc2f6563d6af24bc85953c1ebe",
                "parameters": [
                    {
                        "name": "title_id",
                        "in": "path",
                        "description": "Title ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        },
                        "example": "39535"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/MALEntry"
                                }
                            }
                        }
                    },
                    "401": {
                        "$ref": "#/components/responses/Unauthorized"
                    },
                    "500": {
                        "description": "Scaper Configuration Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/MalScraperConfigErrorResponse"
                                }
                            }
                        }
                    },
                    "503": {
                        "description": "MAL Server Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/MalServerErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "token": []
                    }
                ]
            }
        },
        "/api/mal/{query_string}": {
            "get": {
                "tags": [
                    "MAL"
                ],
                "summary": "Query Titles",
                "operationId": "8134f5865e7f2a3f41bf64cd7a2ad3fd",
                "parameters": [
                    {
                        "name": "query_string",
                        "in": "path",
                        "description": "Title Search String",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "tensei"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/MALSearch"
                                }
                            }
                        }
                    },
                    "401": {
                        "$ref": "#/components/responses/Unauthorized"
                    },
                    "500": {
                        "description": "Scaper Configuration Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/MalScraperConfigErrorResponse"
                                },
                                "examples": {
                                    "ScaperConfigNotFound": {
                                        "$ref": "#/components/examples/ScaperConfigNotFound"
                                    },
                                    "ScraperDisabled": {
                                        "$ref": "#/components/examples/ScraperDisabled"
                                    }
                                }
                            }
                        }
                    },
                    "503": {
                        "description": "MAL Server Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/MalServerErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "token": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "MalScraperConfigErrorResponse": {
                "title": "500 Scraper Configuration Error Responses",
                "properties": {
                    "status": {
                        "type": "string"
                    },
                    "message": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "MalServerErrorResponse": {
                "title": "503 MAL Server Error",
                "properties": {
                    "status": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "message": {
                        "type": "string"
                    }
                },
                "type": "object",
                "example": {
                    "status": "503",
                    "message": "Issues in connecting to MAL Servers"
                }
            },
...
            "Group": {
                "properties": {
                    "uuid": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "name": {
                        "type": "string"
                    }
                },
                "type": "object",
                "example": {
                    "uuid": "e9597119-8452-4f2b-96d8-f2b1b1d2f158",
                    "name": "Sample Group Name"
                }
            },
...
            "MALEntry": {
                "properties": {
                    "url": {
                        "type": "string",
                        "format": "uri"
                    },
                    "title": {
                        "type": "string"
                    },
                    "synonyms": {
                        "type": "string"
                    },
                    "episodes": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "premiered": {
                        "type": "string"
                    }
                },
                "type": "object",
                "example": {
                    "url": "https://example.com/",
                    "title": "Sample Title",
                    "synonyms": "Synonyms",
                    "episodes": 100,
                    "premiered": "Winter 2020"
                }
            },
            "MALSearch": {
                "type": "array",
                "items": {
                    "properties": {
                        "id": {
                            "type": "string"
                        },
                        "title": {
                            "type": "string"
                        }
                    },
                    "type": "object"
                },
                "example": [
                    {
                        "id": "12345",
                        "title": "Sample Title"
                    }
                ]
            },
...
        },
        "responses": {
            "Success": {
                "description": "Success",
                "content": {
                    "application/json": {
                        "schema": {
                            "properties": {
                                "status": {
                                    "type": "integer",
                                    "format": "int32"
                                },
                                "message": {
                                    "type": "string"
                                }
                            },
                            "type": "object"
                        },
                        "example": {
                            "status": 200,
                            "message": "Success"
                        }
                    }
                }
            },
            "Unauthorized": {
                "description": "Unauthorized",
                "content": {
                    "application/json": {
                        "schema": {
                            "properties": {
                                "status": {
                                    "type": "integer",
                                    "format": "int32"
                                },
                                "message": {
                                    "type": "string"
                                }
                            },
                            "type": "object"
                        },
                        "example": {
                            "status": 401,
                            "message": "Unauthorized"
                        }
                    }
                }
            },
...
        },
        "examples": {
            "ScaperConfigNotFound": {
                "summary": "Scaper Configuration Not Found",
                "value": {
                    "status": 500,
                    "message": "Web Scraper configuration not found"
                }
            },
            "ScraperDisabled": {
                "summary": "Scraper Disabled",
                "value": {
                    "status": 500,
                    "message": "Web Scraper is disabled"
                }
            }
        },
        "securitySchemes": {
            "token": {
                "type": "http",
                "description": "Login with email and password to get the authentication token",
                "name": "Authorization",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}